requirements.txt
aiosignal==1.3.2
akshare==1.15.37
amqp==5.2.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backtrader==1.9.78.123
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
bs4==0.0.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
DataRecorder==3.6.2
debugpy==1.8.9
decorator==5.2.0
defusedxml==0.7.1
dnspython==2.7.0
dotenv==0.9.9
DownloadKit==2.0.7
DrissionPage==4.1.0.12
easyquotation==0.7.5
easyutils==0.1.7
efinance==0.5.3
et_xmlfile==2.0.0
eventlet==0.37.0
executing==2.1.0
fake-useragent==1.5.1
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
func_timeout==4.3.5
greenlet==3.1.1
h11==0.14.0
holidays==0.68
html5lib==1.1
httpcore==1.0.7
httpx==0.28.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
json5==0.10.0
jsonpath==0.82.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kombu==5.4.2
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mini-racer==0.12.4
mistune==3.0.2
mkl-service==2.4.0
msgpack==1.1.0
multitasking==0.0.11
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.3.1
notebook_shim==0.2.4
numpy==2.2.3
oauthlib==3.2.2
openpyxl==3.1.5
overrides==7.7.0
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==5.20.0
plotly-express==0.4.1
prettytable==3.10.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
py==1.11.0
py-mini-racer==0.6.0
pydantic==2.10.6
pydantic_core==2.27.2
pydash==7.0.7
pyecharts==2.0.5
PyExecJS==1.5.1
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyquery==2.0.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2025.1
pywencai==0.12.3
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qstock==1.3.6
ray==2.43.0
redis==5.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.1
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.1
textblob==0.18.0.post0
tinycss2==1.4.0
tldextract==5.1.3
tornado==6.4.2
traitlets==5.14.3
tushare==1.4.19
tweepy==4.14.0
types-python-dateutil==2.9.0.20241003
tzdata==2024.1
tzlocal==5.3
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xlrd==2.0.1
xtquant==241014.1.2
yfinance==0.2.54
